Split schemaTypes.spec.ts to match structure of code being tested
